+++ Sources/StreamChat/Utils/Database/NSManagedObject+Extensions.swift
@@ -147,7 +147,7 @@ class FetchCache {
 
 extension NSManagedObjectContext {
     func fetch<T>(_ request: NSFetchRequest<T>, using cache: FetchCache) throws -> [T] where T: NSFetchRequestResult {
-        if let objectIds = cache.get(request) {
+        if let objectIds = cache.get(request), !objectIds.contains(where: { $0.isTemporaryID }) {
             return try objectIds.compactMap { try existingObject(with: $0) as? T }
         }